Move business logic to top level
